Домашнее задание к уроку 6. Добавила возможность общаться с сервером из консоли.
