+++ src/lesson6/client/View.fxml
@@ -1,0 +1,82 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane prefHeight="400.0" prefWidth="337.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="lesson6.client.ViewController">
+    <children>
+        <SplitPane dividerPositions="0.3" prefHeight="400.0" prefWidth="600.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+            <items>
+                <AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
+                    <children>
+                        <ListView fx:id="usersList" minHeight="-Infinity" minWidth="-Infinity" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
+                    </children></AnchorPane>
+                <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
+                    <children>
+                        <SplitPane dividerPositions="0.8838383838383839" orientation="VERTICAL" prefHeight="398.0" prefWidth="486.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+                            <items>
+                                <VBox prefHeight="200.0" prefWidth="100.0">
+                                    <children>
+                                        <HBox alignment="CENTER" maxHeight="-Infinity" prefHeight="30.0" prefWidth="200.0">
+                                            <children>
+                                                <AnchorPane prefHeight="42.0" prefWidth="142.0">
+                                                    <children>
+                                                        <Label prefHeight="42.0" prefWidth="75.0" text="Чат" AnchorPane.leftAnchor="5.0">
+                                                            <font>
+                                                                <Font size="24.0" />
+                                                            </font>
+                                                        </Label>
+                                                    </children>
+                                                </AnchorPane>
+<!--                                                <AnchorPane prefHeight="42.0" prefWidth="275.0">-->
+<!--                                                    <children>-->
+<!--                                                        <Label fx:id="usernameTitle" nodeOrientation="RIGHT_TO_LEFT" prefHeight="42.0" prefWidth="250.0" text="1234" textAlignment="RIGHT" AnchorPane.rightAnchor="12.0">-->
+<!--                                                            <font>-->
+<!--                                                                <Font size="18.0" />-->
+<!--                                                            </font>-->
+<!--                                                        </Label>-->
+<!--                                                    </children>-->
+<!--                                                    <HBox.margin>-->
+<!--                                                        <Insets />-->
+<!--                                                    </HBox.margin>-->
+<!--                                                </AnchorPane>-->
+                                            </children>
+                                        </HBox>
+                                        <TextArea fx:id="chatHistory" editable="false" maxHeight="1.7976931348623157E308" prefWidth="200.0" wrapText="true" VBox.vgrow="ALWAYS">
+                                            <font>
+                                                <Font size="15.0" />
+                                            </font></TextArea>
+                                    </children>
+                                </VBox>
+                                <AnchorPane maxHeight="50.0" minHeight="30.0" minWidth="0.0" prefWidth="160.0">
+                                    <children>
+                                        <HBox prefHeight="100.0" prefWidth="200.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+                                            <children>
+                                                <TextField fx:id="textField" maxHeight="1.7976931348623157E308" prefHeight="43.0" prefWidth="140.0" promptText="Insert your message here" HBox.hgrow="ALWAYS">
+                                                    <font>
+                                                        <Font size="18.0" />
+                                                    </font></TextField>
+                                                <Button fx:id="sendButton" maxHeight="1.7976931348623157E308" mnemonicParsing="false" prefHeight="43.0" prefWidth="66.0" text="Send">
+                                                    <font>
+                                                        <Font size="18.0" />
+                                                    </font></Button>
+                                            </children>
+                                        </HBox>
+                                    </children>
+                                </AnchorPane>
+                            </items>
+                        </SplitPane>
+                    </children>
+                </AnchorPane>
+            </items>
+        </SplitPane>
+    </children>
+</AnchorPane>
